[SourceKit] Add reformat function to libIDE
Move all the rest of the code formatters and helper classes out of the interface.
